+++ src/mesido/esdl/esdl_mixin.py
@@ -79,7 +79,7 @@ class ESDLMixin(
 
         We create a dict with all possible pipe classes for the optional pipes to later add them
         to the optimization problem. Additionally, we change the investment cost figures if an
-        asset pipe template is provided. This is done in this Mixin as we here use the information
+        asset pipe measure is provided. This is done in this Mixin as we here use the information
         of the EDR database which is linked to ESDL and the Mapeditor.
 
         Parameters
@@ -103,7 +103,7 @@ class ESDLMixin(
         self._esdl_assets: Dict[str, Asset] = esdl_parser.get_assets()
         self._esdl_carriers: Dict[str, Dict[str, Any]] = esdl_parser.get_carrier_properties()
         self.__energy_system_handler: esdl.esdl_handler.EnergySystemHandler = esdl_parser.get_esh()
-        self._esdl_templates: Dict[str, Asset] = esdl_parser.get_templates()
+        self._esdl_measures: Dict[str, Asset] = esdl_parser.get_measures()
 
         profile_reader_class = kwargs.get("profile_reader", InfluxDBProfileReader)
         input_file_name = kwargs.get("input_timeseries_file", None)
@@ -200,19 +200,19 @@ class ESDLMixin(
             EDRPipeClass.from_edr_class(name, edr_class_name, maximum_velocity)
             for name, edr_class_name in _AssetToComponentBase.STEEL_S1_PIPE_EDR_ASSETS.items()
         ]
-        # Update the pipe costs if a template model in the ESDL was used. This is updated only if
-        # the pipe catalog is available as a template
-        if self._esdl_templates:
+        # Update the pipe costs if a measure model in the ESDL was used. This is updated only if
+        # the pipe catalog is available as a measure
+        if self._esdl_measures:
             filter_type = "Pipe"
-            pipe_templates = self.filter_asset_templates(
-                asset_templates=self._esdl_templates, filter_type=filter_type
+            pipe_measures = self.filter_asset_measures(
+                asset_measures=self._esdl_measures, filter_type=filter_type
             )
-            if len(pipe_templates.items()) > 0:
+            if len(pipe_measures.items()) > 0:
                 pipe_diameter_cost_map = {
-                    str(pipe.attributes["asset"].diameter): pipe.attributes[
-                        "asset"
+                    str(pipe.attributes["asset"][0].diameter): pipe.attributes["asset"][
+                        0
                     ].costInformation.investmentCosts.value
-                    for pipe in pipe_templates.values()
+                    for pipe in pipe_measures.values()
                 }
                 for i, pipe_class in enumerate(pipe_classes):
                     if pipe_class.name in pipe_diameter_cost_map.keys():
@@ -678,12 +678,12 @@ class ESDLMixin(
         self.__timeseries_export.write()
 
     @classmethod
-    def filter_asset_templates(
-        cls, asset_templates: Dict[str, Asset], filter_type: str
+    def filter_asset_measures(
+        cls, asset_measures: Dict[str, Asset], filter_type: str
     ) -> Dict[str, Asset]:
         filtered_assets = dict()
-        for asset_id, asset in asset_templates.items():
-            asset_type = asset.attributes["asset"]
+        for asset_id, asset in asset_measures.items():
+            asset_type = asset.attributes["asset"][0]
             if isinstance(asset_type, getattr(esdl, filter_type)):
                 filtered_assets[asset_id] = asset